==> MF_FYP/25_dinucleotide_hist.R <==
#comparing observed vs expected dinucleotide rates in genomes 
library(ggplot2)
  

pdf("25.1_fig_5.pdf")


data_file <- read.csv('11.1_dinucleotide_content.csv',header = TRUE)

predicted_CT_all <- (unlist(data_file$predicted_CT))
predicted_CG_all <- (unlist(data_file$predicted_CG))
predicted_TG_all <- (unlist(data_file$predicted_TG))
predicted_CA_all <- (unlist(data_file$predicted_CT))
predicted_TT_all <- (unlist(data_file$predicted_TT))
predicted_TA_all <- (unlist(data_file$predicted_TA))
predicted_GG_all <- (unlist(data_file$predicted_GG))
predicted_AG_all <- (unlist(data_file$predicted_AG))
predicted_CC_all <- (unlist(data_file$predicted_CC))
predicted_GT_all <- (unlist(data_file$predicted_GT))
predicted_GA_all <- (unlist(data_file$predicted_GA))
predicted_AT_all <- (unlist(data_file$predicted_AT))
predicted_TC_all <- (unlist(data_file$predicted_TC))
predicted_AA_all <- (unlist(data_file$predicted_AA))
predicted_GC_all<- (unlist(data_file$predicted_GC))
predicted_AC_all <- (unlist(data_file$predicted_AC))


actual_CT_all <- (unlist(data_file$actual_CT))
actual_CG_all <- (unlist(data_file$actual_CG))
actual_TG_all <- (unlist(data_file$actual_TG))
actual_CA_all <- (unlist(data_file$actual_CT))
actual_TT_all <- (unlist(data_file$actual_TT))
actual_TA_all <- (unlist(data_file$actual_TA))
actual_GG_all <- (unlist(data_file$actual_GG))
actual_AG_all <- (unlist(data_file$actual_AG))
actual_CC_all <- (unlist(data_file$actual_CC))
actual_GT_all <- (unlist(data_file$actual_GT))
actual_GA_all <- (unlist(data_file$actual_GA))
actual_AT_all <- (unlist(data_file$actual_AT))
actual_TC_all <- (unlist(data_file$actual_TC))
actual_AA_all <- (unlist(data_file$actual_AA))
actual_GC_all <- (unlist(data_file$actual_GC))
actual_AC_all <- (unlist(data_file$actual_AC))



predicted_CT <- mean(unlist(data_file$predicted_CT))
predicted_CG <- mean(unlist(data_file$predicted_CG))
predicted_TG <- mean(unlist(data_file$predicted_TG))
predicted_CA <- mean(unlist(data_file$predicted_CT))
predicted_TT <- mean(unlist(data_file$predicted_TT))
predicted_TA <- mean(unlist(data_file$predicted_TA))
predicted_GG <- mean(unlist(data_file$predicted_GG))
predicted_AG <- mean(unlist(data_file$predicted_AG))
predicted_CC <- mean(unlist(data_file$predicted_CC))
predicted_GT <- mean(unlist(data_file$predicted_GT))
predicted_GA <- mean(unlist(data_file$predicted_GA))
predicted_AT <- mean(unlist(data_file$predicted_AT))
predicted_TC <- mean(unlist(data_file$predicted_TC))
predicted_AA <- mean(unlist(data_file$predicted_AA))
predicted_GC <- mean(unlist(data_file$predicted_GC))
predicted_AC <- mean(unlist(data_file$predicted_AC))


actual_CT <- mean(unlist(data_file$actual_CT))
actual_CG <- mean(unlist(data_file$actual_CG))
actual_TG <- mean(unlist(data_file$actual_TG))
actual_CA <- mean(unlist(data_file$actual_CT))
actual_TT <- mean(unlist(data_file$actual_TT))
actual_TA <- mean(unlist(data_file$actual_TA))
actual_GG <- mean(unlist(data_file$actual_GG))
actual_AG <- mean(unlist(data_file$actual_AG))
actual_CC <- mean(unlist(data_file$actual_CC))
actual_GT <- mean(unlist(data_file$actual_GT))
actual_GA <- mean(unlist(data_file$actual_GA))
actual_AT <- mean(unlist(data_file$actual_AT))
actual_TC <- mean(unlist(data_file$actual_TC))
actual_AA <- mean(unlist(data_file$actual_AA))
actual_GC <- mean(unlist(data_file$actual_GC))
actual_AC <- mean(unlist(data_file$actual_AC))


x <- c('CT','GC','TG','CA','TT','TA','GG','AG','CC','GT','GA','AT','TC','AA','GC','AC')
CT=c(predicted_CT, actual_CT)
CG=c(predicted_CG, actual_CG)
TG=c(predicted_TG, actual_TG)
CA=c(predicted_CA, actual_CA)
TT=c(predicted_TT, actual_TT)
TA=c(predicted_TA, actual_TA)
GG=c(predicted_GG, actual_GG)
AG=c(predicted_AG, actual_AG)
CC=c(predicted_CC, actual_CC)
GT=c(predicted_GT, actual_GT)
GA=c(predicted_GA, actual_GA)
AT=c(predicted_AT, actual_AT)
TC=c(predicted_TC, actual_TC)
AA=c(predicted_AA, actual_AA)
GC=c(predicted_GC, actual_GC)
AC=c(predicted_AC, actual_AC)

# Inserting data
data <- data.frame(CT,CG,TG,CA,TT,TA,GG,AG,CC,GT,GA,AT,TC,AA,GC,AC)
colour_dict <- c("turquoise3","plum")

barplot(as.matrix(data), xlab="mean frequency",ylab = "dinucleotide", horiz=T, col = colour_dict,width = 3, legend = c("predicted","actual"), xlim = c(0.00,0.1), beside = TRUE, cex.names=0.7, cex.axis = 0.8, args.legend = list(x="right"))

#stats 
print(paste('CT'))
stats_CT <- t.test(actual_CT_all,predicted_CT_all, paired = TRUE)
print(stats_CT)
print(paste('CG'))
stats_CG <- t.test(actual_CG_all,predicted_CG_all, paired = TRUE)
print(stats_CG)
print(paste('TG'))
stats_TG <- t.test(actual_TG_all,predicted_TG_all, paired = TRUE)
print(stats_TG)
print(paste('CA'))
stats_CA <- t.test(actual_CA_all,predicted_CA_all, paired = TRUE)
print(stats_CA)
print(paste('TT'))
stats_TT <- t.test(actual_TT_all,predicted_TT_all, paired = TRUE)
print(stats_TT)
print(paste('TA'))
stats_TA <- t.test(actual_TA_all,predicted_TA_all, paired = TRUE)
print(stats_TA)
print(paste('GG'))
stats_GG <- t.test(actual_GG_all,predicted_GG_all, paired = TRUE)
print(stats_GG)
print(paste('AG'))
stats_AG <- t.test(actual_AG_all,predicted_AG_all, paired = TRUE)
print(stats_AG)
print(paste('CC'))
stats_CC <- t.test(actual_CC_all,predicted_CC_all, paired = TRUE)
print(stats_CC)
print(paste('GT'))
stats_GT <- t.test(actual_GT_all,predicted_GT_all, paired = TRUE)
print(stats_GT)
print(paste('GA'))
stats_GA <- t.test(actual_GA_all,predicted_GA_all, paired = TRUE)
print(stats_GA)
print(paste('AT'))
stats_AT <- t.test(actual_AT_all,predicted_AT_all, paired = TRUE)
print(stats_AT)
print(paste('TC'))
stats_TC <- t.test(actual_TC_all,predicted_TC_all, paired = TRUE)
print(stats_TC)
print(paste('AA'))
stats_AA <- t.test(actual_AA_all,predicted_AA_all, paired = TRUE)
print(stats_AA)
print(paste('GC'))
stats_GC <- t.test(actual_GC_all,predicted_GC_all, paired = TRUE)
print(stats_GC)
print(paste('AC'))
stats_AC <- t.test(actual_AC_all,predicted_AC_all, paired = TRUE)
print(stats_AC)

dev.off()
